Introduce Proofing Mode with MinWidth and MinHeight
Proofing mode is a boolean flag that determines if we use a special fallback
image that is more obvious. When preparing an image for production use,
the fallback might not be obvious. With proofing mode, (proof, pf), a
special fallback image can be provided to make this more obvious. Then
when the image is correctly prepared it can then be used in production.
Add Proof bool parameter which determines if we use proof fallback or
normal fallback.

Part of proofing mode is also the concept of MinWidth and MinHeight.
These are bounds for how small an image can be before the fallback kicks
in as the image would look too terrible for actual use otherwise.